requirements.txt
aiogram==2.25.1
aiofiles==24.1.0
apscheduler==3.10.4
aiohappyeyeballs==2.6.1
aiohttp==3.8.5
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
Django==5.2.8
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
redis==7.1.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
celery
redis
django_celery_beat
